Fix type definitions and remaining lint errors
