+++ error-issue/src/blame_finder.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {BlameRange, GraphQLResponse, ILogger, StackFrame} from './types';
+import {BlameRange, GraphQLResponse, Logger, StackFrame} from 'error-issue-bot';
 import {RateLimitedGraphQL} from './rate_limited_graphql';
 import {parsePrNumber, parseStacktrace} from './utils';
 
@@ -30,7 +30,7 @@ export class BlameFinder {
     private repoOwner: string,
     private repoName: string,
     client: RateLimitedGraphQL,
-    private logger: ILogger = console
+    private logger: Logger = console
   ) {
     this.graphql = async (query: string): Promise<GraphQLResponse> =>
       client.runQuery(query);
