+++ error-issue/src/rate_limited_graphql.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {GraphQLResponse} from './types';
+import {GraphQLResponse} from 'error-issue-bot';
 import {graphql} from '@octokit/graphql';
 
 const GRAPHQL_FREQ_MS = parseInt(process.env.GRAPHQL_FREQ_MS, 10) || 100;
