+++ error-issue/src/utils.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {StackFrame} from './types';
+import {StackFrame} from 'error-issue-bot';
 
 const SOURCE_PREFIX = 'https://raw.githubusercontent.com/ampproject/amphtml/';
 const SOURCE_PATTERN = /^(?<rtv>\d+)\/(?<path>[^:]+):(?<line>\d+)(?<column>:\d+)?$/;
