+++ error-issue/test/issue_builder.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {BlameRange, ErrorReport} from '../src/types';
+import {BlameRange, ErrorReport} from 'error-issue-bot';
 import {IssueBuilder} from '../src/issue_builder';
 import {getFixtureFile} from './fixtures';
