+++ error-issue/tsconfig.json
@@ -11,7 +11,7 @@
     "baseUrl": ".",
     "paths": {
       "*": [
-        "node_modules/*",        
+        "node_modules/*",
         "node_modules/@types/*",
         "src/types/*"
       ]
@@ -20,6 +20,7 @@
   "include": [
     "./*",
     "src/**/*",
-    "test/*"
+    "test/*",
+    "types/*"
   ]
 }
